+++ src/main/java/de/dm/intellij/liferay/language/freemarker/LiferayFreemarkerTaglibDeprecationInspection.java
@@ -70,6 +70,8 @@ import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_166546_AUI
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_166546_LIFERAY_AUI;
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_168309_LIFERAY_AUI;
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_168309_LIFERAY_FRONTEND;
+import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_199170_COMMERCE;
+import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_199170_EXPORT_IMPORT_CHANGESET;
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_54620_PORTLET_ICON;
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_55886_APP_VIEW_SEARCH_ENTRY;
 import static de.dm.intellij.liferay.util.LiferayTaglibAttributes.LPS_60328_NAVIGATION;
@@ -168,6 +170,8 @@ public class LiferayFreemarkerTaglibDeprecationInspection extends AbstractLifera
 		TAGLIB_DEPRECATIONS.addAll(createTags(LPS_158461_LIFERAY_FRONTEND).quickfix(removeTag()));
 		TAGLIB_DEPRECATIONS.addAll(createTags(LPS_166546_AUI_CONTAINER).quickfix(renameNamespace(TAGLIB_URI_LIFERAY_CLAY)));
 		TAGLIB_DEPRECATIONS.addAll(createTags(LPS_166546_LIFERAY_AUI).quickfix(removeTag()));
+		TAGLIB_DEPRECATIONS.addAll(createTags(LPS_199170_COMMERCE).quickfix(removeTag()).version("7.4.3.100"));
+		TAGLIB_DEPRECATIONS.addAll(createTags(LPS_199170_EXPORT_IMPORT_CHANGESET).quickfix(removeTag()).version("7.4.3.100"));
 	}
 
 	@Nls